Moves State-Engine execution to context phase.
The reason why it happened on both the input and output phases, but not the context phase, has to do with the edit distance limitation that AI Dungeon used to have.  It would inject a pre-calculated `state.memory.context` to reduce the chances that the edit distance constraint would be triggered when the context phase came around and Context-Mode started to make edits.
Since this constraint is gone, we don't need to put up with this unnecessary complication any more.